Re-license RITlug Runbook as CC-BY-SA 4.0 + MPL 2.0
This commit sets the licensing of the RITlug Runbook to a dual-license
project.

Technical contributions
=======================

Code, scripts, or any machine-read content is licensed under the
Mozilla Public License 2.0.

In the event any serious code or technical content is added to the
Runbook, it will be the handiwork of one of a RITlug eboard member or
community member. The MPL is chosen and selected to promise that the
project built here will remain open source and anyone who makes changes
to our Runbook code must also open source their changes. This is a help
to our members to make their work more visible if reused.

Actual Runbook pages / content
==============================

Most of the content in this repo is not code or scripts, but written
guidance and instruction. This content is valuable and interesting, and
how we set up our Runbook is unique. We want to share our content with
the wider open source community but also maintain that anyone reusing
our content or using it to build their own documentation must
distribute derivative works under an open license and retain
attribution to the RITlug Runbook.

This is also to benefit the publicity of work done by students and to
encourage a positive example implemented by other communities.

Approval
========

This pull request requires approval by all four of the Runbook's
contributors to be carried. Those four people, by GitHub username,
are:

* @jflory7: +1
* @thenaterhood
* @Serubin
* @axk4545

Change discussed and reviewed:
  https://github.com/RITlug/runbook/pull/31